requirements.txt
fastapi
uvicorn
PyPDF2
tqdm
langchain-openai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
openai
